rag-server/requirements.txt
chromadb>=0.4.15
sentence-transformers>=2.2.2
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
sqlalchemy
alembic
psycopg2-binary
python-dotenv
